work in progress
Former-commit-id: 73e4bd633bfec0c98adcdee98a1f0f7c174fee56
Former-commit-id: c325ea0173a8f029529cab77bf6f06faddf33299

+++ src/wid_game_map_client.h
@@ -18,7 +18,7 @@ void wid_game_map_client_scroll_adjust(void);
 boolean wid_game_map_client_player_move(void);
 
 extern widp
-wid_game_map_client_replace_tile(widp, int32_t x, int32_t y, thingp t);
+wid_game_map_client_replace_tile(widp, double x, double y, thingp t);
 
 extern widp wid_game_map_client_window;
 extern widp wid_game_map_client_grid_container;
